+++ packages/attach/src/component.screenshots.test.tsx
@@ -25,7 +25,10 @@ describe('Attach', () =>
                 noFileText: ' ',
             },
         },
-        'paddingLeft:3px;transform: scale(2.1)',
+        'width:800px;paddingLeft:3px;transform: scale(2.1)',
+        {
+            viewport: { width: 1024, height: 600 },
+        },
     ));
 
 describe(