+++ packages/checkbox-group/src/component.screenshots.test.tsx
@@ -21,7 +21,10 @@ describe('CheckboxGroup', () => {
                 checked: true,
             },
         },
-        'transform:scale(2.3)',
+        'width:800px;transform:scale(2.3)',
+        {
+            viewport: { width: 1024, height: 600 },
+        },
     );
 });