+++ packages/date-time-input/src/components/date-time-input/Component.tsx
@@ -13,14 +13,11 @@ import React, {
 import mergeRefs from 'react-merge-refs';
 import cn from 'classnames';
 
-import {
-    Calendar as DefaultCalendar,
-    CalendarMobileProps,
-    CalendarProps,
-    dateInLimits,
-} from '@alfalab/core-components-calendar';
+import type { CalendarDesktopProps } from '@alfalab/core-components-calendar/desktop';
+import type { CalendarMobileProps } from '@alfalab/core-components-calendar/mobile';
+import { dateInLimits } from '@alfalab/core-components-calendar/shared';
 import { IconButton } from '@alfalab/core-components-icon-button';
-import { Input, InputProps } from '@alfalab/core-components-input';
+import { InputProps } from '@alfalab/core-components-input';
 import { Popover, PopoverProps } from '@alfalab/core-components-popover';
 import { useDidUpdateEffect } from '@alfalab/hooks';
 import { CalendarMIcon } from '@alfalab/icons-glyph/CalendarMIcon';
@@ -84,7 +81,7 @@ export type DateTimeInputProps = Omit<InputProps, 'onChange'> & {
      * Доп. пропсы для календаря
      */
     calendarProps?:
-        | (CalendarProps & Record<string, unknown>)
+        | (CalendarDesktopProps & Record<string, unknown>)
         | (CalendarMobileProps & Record<string, unknown>);
 
     /**
@@ -149,6 +146,11 @@ export type DateTimeInputProps = Omit<InputProps, 'onChange'> & {
     view?: 'desktop' | 'mobile';
 
     /**
+     * Компонент инпута
+     */
+    InputComponent?: ElementType;
+
+    /**
      * Запретить ввод с клавиатуры
      */
     disableUserInput?: boolean;
@@ -174,7 +176,8 @@ export const DateTimeInput = React.forwardRef<HTMLInputElement, DateTimeInputPro
             popoverPosition = 'bottom-start',
             zIndexPopover,
             preventFlip,
-            Calendar = DefaultCalendar,
+            InputComponent,
+            Calendar,
             calendarProps = {},
             defaultMonth,
             minDate = calendarProps.minDate,
@@ -339,25 +342,26 @@ export const DateTimeInput = React.forwardRef<HTMLInputElement, DateTimeInputPro
             }
         };
 
-        const renderCalendar = () => (
-            // eslint-disable-next-line jsx-a11y/no-static-element-interactions
-            <div onMouseDown={handleCalendarWrapperMouseDown}>
-                <Calendar
-                    {...calendarProps}
-                    responsive={calendarResponsive}
-                    open={open}
-                    onClose={handleMobileCalendarClose}
-                    ref={calendarRef}
-                    defaultMonth={defaultMonth}
-                    value={checkInputValueIsValid(value) ? calendarValue : undefined}
-                    onChange={handleCalendarChange}
-                    minDate={minDate}
-                    maxDate={maxDate}
-                    offDays={offDays}
-                    events={events}
-                />
-            </div>
-        );
+        const renderCalendar = () =>
+            Calendar ? (
+                // eslint-disable-next-line jsx-a11y/no-static-element-interactions
+                <div onMouseDown={handleCalendarWrapperMouseDown}>
+                    <Calendar
+                        {...calendarProps}
+                        responsive={calendarResponsive}
+                        open={open}
+                        onClose={handleMobileCalendarClose}
+                        ref={calendarRef}
+                        defaultMonth={defaultMonth}
+                        value={checkInputValueIsValid(value) ? calendarValue : undefined}
+                        onChange={handleCalendarChange}
+                        minDate={minDate}
+                        maxDate={maxDate}
+                        offDays={offDays}
+                        events={events}
+                    />
+                </div>
+            ) : null;
 
         return (
             <div
@@ -367,32 +371,34 @@ export const DateTimeInput = React.forwardRef<HTMLInputElement, DateTimeInputPro
                 onFocus={inputDisabled ? undefined : handleInputWrapperFocus}
                 onBlur={handleBlur}
             >
-                <Input
-                    {...restProps}
-                    block={block}
-                    ref={mergeRefs([ref, inputRef])}
-                    value={value}
-                    onChange={handleChange}
-                    disabled={disabled}
-                    readOnly={readOnly}
-                    className={inputClassName}
-                    onClear={handleClear}
-                    onKeyDown={handleInputKeyDown}
-                    error={error}
-                    rightAddons={
-                        <React.Fragment>
-                            {rightAddons}
-                            {picker && (
-                                <IconButton
-                                    className={styles.calendarIcon}
-                                    onClick={inputDisabled ? undefined : handleIconButtonClick}
-                                    icon={CalendarMIcon}
-                                    size='s'
-                                />
-                            )}
-                        </React.Fragment>
-                    }
-                />
+                {InputComponent ? (
+                    <InputComponent
+                        {...restProps}
+                        block={block}
+                        ref={mergeRefs([ref, inputRef])}
+                        value={value}
+                        onChange={handleChange}
+                        disabled={disabled}
+                        readOnly={readOnly}
+                        className={inputClassName}
+                        onClear={handleClear}
+                        onKeyDown={handleInputKeyDown}
+                        error={error}
+                        rightAddons={
+                            <React.Fragment>
+                                {rightAddons}
+                                {picker && (
+                                    <IconButton
+                                        className={styles.calendarIcon}
+                                        onClick={inputDisabled ? undefined : handleIconButtonClick}
+                                        icon={CalendarMIcon}
+                                        size='s'
+                                    />
+                                )}
+                            </React.Fragment>
+                        }
+                    />
+                ) : null}
                 {picker && (
                     <Popover
                         open={open}