+++ packages/modal/src/mobile.ts
@@ -1,1 +1,2 @@
 export * from './Component.mobile';
+export { ModalMobileProps } from './typings';