+++ packages/number-input/src/Component.screenshots.test.tsx
@@ -11,5 +11,8 @@ describe('NumberInput ', () =>
                 block: true,
             },
         },
-        'padding: 0 270px; transform:scale(2.1)',
+        'padding: 0 270px;width:800px;transform:scale(2.1)',
+        {
+            viewport: { width: 1024, height: 600 },
+        },
     ));