+++ packages/toast-plate/src/component.test.tsx
@@ -1,11 +1,25 @@
 import React from 'react';
 import { fireEvent, render } from '@testing-library/react';
 
-import { ToastPlate } from './index';
+import { ToastPlateDesktop as ToastPlate } from './desktop';
 
 jest.useFakeTimers();
 
 describe('Notification', () => {
+    Object.defineProperty(window, 'matchMedia', {
+        writable: true,
+        value: jest.fn().mockImplementation((query) => ({
+            matches: true,
+            media: query,
+            onchange: null,
+            addListener: jest.fn(), // Deprecated
+            removeListener: jest.fn(), // Deprecated
+            addEventListener: jest.fn(),
+            removeEventListener: jest.fn(),
+            dispatchEvent: jest.fn(),
+        })),
+    });
+
     describe('Snapshots tests', () => {
         it('should match snapshot', () => {
             const { baseElement } = render(