+++ packages/tooltip/src/desktop.ts
@@ -1,1 +1,2 @@
-export * from './Component';
+export { TooltipDesktop } from './Component.desktop';
+export { TooltipDesktopProps } from './types';